ai/requirements.txt
pandas
matplotlib
statsmodels
psycopg2-binary
gunicorn
fpdf
flask